+++ ray_tracer/src/rays.rs
@@ -22,7 +22,7 @@ impl Ray {
     pub fn get_direction(&self) -> Vector {
         self.direction
     }
-    pub fn position(&self, time: f32) -> Point {
+    pub fn position(&self, time: f64) -> Point {
         self.origin + self.direction * time
     }
