+++ ray_tracer/src/transformations.rs
@@ -3,7 +3,7 @@ use crate::{matrices::Matrix, tuples::Tuple};
 #[derive(Debug, PartialEq, Clone, Copy)]
 pub struct Transform;
 impl Transform {
-    pub fn translate(x: f32, y: f32, z: f32) -> Matrix {
+    pub fn translate(x: f64, y: f64, z: f64) -> Matrix {
         Matrix::new(vec![
             vec![1.0, 0.0, 0.0, x],
             vec![0.0, 1.0, 0.0, y],
@@ -12,7 +12,7 @@ impl Transform {
         ])
         .unwrap()
     }
-    pub fn scaling(x: f32, y: f32, z: f32) -> Matrix {
+    pub fn scaling(x: f64, y: f64, z: f64) -> Matrix {
         Matrix::new(vec![
             vec![x, 0.0, 0.0, 0.0],
             vec![0.0, y, 0.0, 0.0],
@@ -21,34 +21,34 @@ impl Transform {
         ])
         .unwrap()
     }
-    pub fn rotation_x(angle: f32) -> Matrix {
+    pub fn rotation_x(angle: f64) -> Matrix {
         Matrix::new(vec![
             vec![1.0, 0.0, 0.0, 0.0],
-            vec![0.0, f32::cos(angle), -f32::sin(angle), 0.0],
-            vec![0.0, f32::sin(angle), f32::cos(angle), 0.0],
+            vec![0.0, f64::cos(angle), -f64::sin(angle), 0.0],
+            vec![0.0, f64::sin(angle), f64::cos(angle), 0.0],
             vec![0.0, 0.0, 0.0, 1.0],
         ])
         .unwrap()
     }
-    pub fn rotation_y(angle: f32) -> Matrix {
+    pub fn rotation_y(angle: f64) -> Matrix {
         Matrix::new(vec![
-            vec![f32::cos(angle), 0.0, f32::sin(angle), 0.0],
+            vec![f64::cos(angle), 0.0, f64::sin(angle), 0.0],
             vec![0.0, 1.0, 0.0, 0.0],
-            vec![-f32::sin(angle), 0.0, f32::cos(angle), 0.0],
+            vec![-f64::sin(angle), 0.0, f64::cos(angle), 0.0],
             vec![0.0, 0.0, 0.0, 1.0],
         ])
         .unwrap()
     }
-    pub fn rotation_z(angle: f32) -> Matrix {
+    pub fn rotation_z(angle: f64) -> Matrix {
         Matrix::new(vec![
-            vec![f32::cos(angle), -f32::sin(angle), 0.0, 0.0],
-            vec![f32::sin(angle), f32::cos(angle), 0.0, 0.0],
+            vec![f64::cos(angle), -f64::sin(angle), 0.0, 0.0],
+            vec![f64::sin(angle), f64::cos(angle), 0.0, 0.0],
             vec![0.0, 0.0, 1.0, 0.0],
             vec![0.0, 0.0, 0.0, 1.0],
         ])
         .unwrap()
     }
-    pub fn shearing(x_y: f32, x_z: f32, y_x: f32, y_z: f32, z_x: f32, z_y: f32) -> Matrix {
+    pub fn shearing(x_y: f64, x_z: f64, y_x: f64, y_z: f64, z_x: f64, z_y: f64) -> Matrix {
         Matrix::new(vec![
             vec![1.0, x_y, x_z, 0.0],
             vec![y_x, 1.0, y_z, 0.0],
@@ -77,7 +77,7 @@ impl Transform {
 
 #[cfg(test)]
 mod tests {
-    use std::f32::consts::PI;
+    use std::f64::consts::PI;
 
     use super::*;
     use crate::tuples::Tuple;
@@ -148,7 +148,7 @@ mod tests {
 
         assert_eq!(
             half_quarter * p,
-            Tuple::new_point(0.0, f32::sqrt(2.0) / 2.0, f32::sqrt(2.0) / 2.0)
+            Tuple::new_point(0.0, f64::sqrt(2.0) / 2.0, f64::sqrt(2.0) / 2.0)
         );
         assert_eq!(full_quarter * p, Tuple::new_point(0.0, 0.0, 1.0));
     }
@@ -161,7 +161,7 @@ mod tests {
 
         assert_eq!(
             inv * p,
-            Tuple::new_point(0.0, f32::sqrt(2.0) / 2.0, -(f32::sqrt(2.0) / 2.0))
+            Tuple::new_point(0.0, f64::sqrt(2.0) / 2.0, -(f64::sqrt(2.0) / 2.0))
         );
     }
 
@@ -173,7 +173,7 @@ mod tests {
 
         assert_eq!(
             half_quarter * p,
-            Tuple::new_point(f32::sqrt(2.0) / 2.0, 0.0, f32::sqrt(2.0) / 2.0)
+            Tuple::new_point(f64::sqrt(2.0) / 2.0, 0.0, f64::sqrt(2.0) / 2.0)
         );
         assert_eq!(full_quarter * p, Tuple::new_point(1.0, 0.0, 0.0));
     }
@@ -186,7 +186,7 @@ mod tests {
 
         assert_eq!(
             half_quarter * p,
-            Tuple::new_point(-(f32::sqrt(2.0) / 2.0), f32::sqrt(2.0) / 2.0, 0.0)
+            Tuple::new_point(-(f64::sqrt(2.0) / 2.0), f64::sqrt(2.0) / 2.0, 0.0)
         );
         assert_eq!(full_quarter * p, Tuple::new_point(-1.0, 0.0, 0.0));
     }
