+++ ray_tracer/src/utils.rs
@@ -1,8 +1,8 @@
-/// A fixed value used for comparing f32
-pub const EPSILON: f32 = 0.00005;
+/// A fixed value used for comparing f64
+pub const EPSILON: f64 = 0.00005;
 
-/// Compares two f32 and asserts whether they are "equal" or not.
-pub fn is_float_equal(actual: &f32, comparison: f32) -> bool {
+/// Compares two f64 and asserts whether they are "equal" or not.
+pub fn is_float_equal(actual: &f64, comparison: f64) -> bool {
     if (actual - comparison).abs() < EPSILON {
         true
     } else {
