+++ src/ch6_pit/ch6_pit.rs
@@ -1,4 +1,4 @@
-use std::{f32::consts::PI, ops::Deref};
+use std::{f64::consts::PI, ops::Deref};
 
 use ray_tracer::{
     canvas::Canvas,
@@ -30,10 +30,10 @@ fn main() {
     squash.set_transform(&transform);
 
     let ray_origin = Tuple::new_point(0.0, 0.0, -5.0);
-    let wall_z = 10.0_f32;
-    let wall_size = 7.0_f32;
+    let wall_z = 10.0_f64;
+    let wall_size = 7.0_f64;
     let canvas_pixels = 100;
-    let pixel_size = wall_size / canvas_pixels as f32;
+    let pixel_size = wall_size / canvas_pixels as f64;
     let half = wall_size / 2.0;
     let mut img = Canvas::new(canvas_pixels, canvas_pixels);
 
@@ -41,12 +41,12 @@ fn main() {
         // Calculate "world y coordinate"
         // Top = +half
         // Bottom = -half
-        let world_y = half - pixel_size * y as f32;
+        let world_y = half - pixel_size * y as f64;
         for x in 0..canvas_pixels {
             // Calculate "world x coordinate"
             // Left = -half
             // Right = +half
-            let world_x = -half + pixel_size * x as f32;
+            let world_x = -half + pixel_size * x as f64;
 
             // Describe the "wall" location the ray will target
             let position = Tuple::new_point(world_x, world_y, wall_z);
